contracts-bedrock: prevent too low of gas config
The chain liveliness depends on the L2 gas limit being larger
than the gas consumed by a deposit. If a deposit consumes more
gas than what is possible in an L2 block, then the chain will
halt and be unable to produce blocks.

This was previously handled by having a hardcoded constant in
the system config contract. This constant must be kept in sync
with the value in the optimism portal. It is also possible to
deploy a new optimism portal such that it has a different resource
limit. To remove the need to always having to keep these values in
sync, a handle to the portal is given to the system config contract.
The system config contract will call the portal to get the value
dynamically.

This ensures that we do not need to remember what the value is
and update the contracts in tandem.